face_landmark_model/requirements.txt
bentoml==1.4.0
numpy==1.24.3
opencv-python==4.8.0.76
mediapipe==0.10.8
Pillow==10.0.0
annotated-types==0.6.0 